Draw hexagon tiles

Draw a grid of tessellated hexagons over the bounding box of a SpatialPolygons object

hex_tiles(x, cellsize = NULL)

Arguments

x
An sp object

cellsize
The size of the hexagons, if left blank then will take the average area of the polygons in the SpatialPolygons data.frame

Value

A SpatialPolygonsDataFrame of tessellated hexagons covering the bounding box of a SpatialPolygons

Examples

require(rworldmap);

#> Loading required package: rworldmap

#> Loading required package: sp

#> ### Welcome to rworldmap ###

#> For a short introduction type : 	 vignette('rworldmap')

data("countriesCoarseLessIslands") #  Load simple map without islands
afr <- countriesCoarseLessIslands[which(!is.na(countriesCoarseLessIslands@data$REGION) &
                                         countriesCoarseLessIslands@data$REGION=="Africa"),]
afr <- sp::spTransform(afr, CRS("+init=EPSG:32663")) # Project to equidistant grid
plot(hex_tiles(afr)[afr,]) # Clip to original shape and plot